alert-checker/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
slack-sdk==3.17.1
requests==2.28.0

azure-functions
